+++ lib/stubberry.rb
@@ -1,8 +1,8 @@
 require "stubberry/version"
 require "stubberry/object"
 require "stubberry/active_record"
+require "stubberry/assertions"
 
 module Stubberry
   class Error < StandardError; end
-  # Your code goes here...
 end