+++ lib/stubberry/active_record.rb
@@ -14,16 +14,16 @@ module Stubberry::ActiveRecord
   # we adding after_find callback extending object with self.class.extend_any
   # default implementation of self.class.extend_any does nothing
   included do
-    after_find {|obj| self.class.extend_any(obj) }
+    after_find {|obj| self.class.__extend_any(obj) }
   end
 
   module ClassMethods
 
     # This method could be used whenever there is a need for stubbing
     # the exact ActiveRecord object attributes inside some execution flow
-    # __without__ underlying record change
+    # __WITHOUT__!!  underlying record change
     def stub_orm_attr(id, obj_or_attributes )
-      stub(:extend_any, -> (obj) {
+      stub(:__extend_any, -> (obj) {
         return unless obj.id == id && obj.is_a?( self )
         obj.assign_attributes( obj_or_attributes.try(:attributes) || obj_or_attributes )
       }) do
@@ -33,23 +33,27 @@ module Stubberry::ActiveRecord
 
     # This method could be used whenever there is a need for stubbing
     # the specific Active Record object's methods inside some flow piece
-    # with ANY way of object retrieval, making
+    # with ANY way of object retrieval
     def stub_orm_method(id, method, val_or_callable, *block_args  )
-      stub(:extend_any, -> (obj) {
+      stub(:__extend_any, -> (obj) {
         return unless obj.id == id && obj.is_a?( self )
-        define_stub_method(obj, method, val_or_callable, *block_args )
+        __define_stub_method(obj, method, val_or_callable, *block_args )
       }) do
         yield
       end
     ensure
-      undef_all(id, method)
+      __revert_all_methods(id, method)
     end
 
     private_class_method
 
-    def define_stub_method( object, method, val_or_callable, *block_args )
-      old_method = stub_method_name(method, obj: object)
-      object.singleton_class.send :alias_method, old_method, method
+    def __define_stub_method(object, method, val_or_callable, *block_args )
+      method_new_name = __stub_method_name(method, obj: object)
+
+      __define_dynamic_method_replacement( object, method )
+
+      object.singleton_class.alias_method( method_new_name, method )
+      
       object.define_singleton_method method do |*args, &blk|
         if val_or_callable.respond_to? :call
           val_or_callable.call(*args, &blk)
@@ -58,27 +62,35 @@ module Stubberry::ActiveRecord
           val_or_callable
         end
       end
-      stubbed_objects(old_method) << object
+      __stubbed_objects(method_new_name) << object
     end
 
-    def stub_method_name( method, obj: nil, id: nil)
+    def __define_dynamic_method_replacement( object, method )
+      # this means method is dynamic 
+      return unless object.respond_to?( method ) && !object.methods.map(&:to_s).include?( method.to_s )
+
+      object.define_singleton_method( method ) { | *args, **kargs, &block | super(*args, **kargs, &block) }
+    end
+
+    def __stub_method_name(method, obj: nil, id: nil)
       # __stub_class_method_id
       "__stub_#{to_s}_#{method}_#{obj&.id || id}"
     end
 
-    def undef_all(id, method)
-      old_method = stub_method_name( method, id: id)
-      stubbed_objects(old_method).map(&:singleton_class).each do |metaclass|
+    def __revert_all_methods(id, method)
+      method_new_name = __stub_method_name(method, id: id)
+
+      __stubbed_objects(method_new_name).map(&:singleton_class).each do |metaclass|
         metaclass.send :undef_method, method
-        metaclass.send :alias_method, method, old_method
-        metaclass.send :undef_method, old_method
+        metaclass.send :alias_method, method, method_new_name
+        metaclass.send :undef_method, method_new_name
       end
     end
 
-    def stubbed_objects(method_name)
-      (@@extended_obj ||= {})[method_name] ||= []
+    def __stubbed_objects(method_name)
+      (@@__extended_objects ||= {})[method_name] ||= []
     end
 
-    def extend_any(_obj); :do_nothing end
+    def __extend_any(_obj); :do_nothing end
   end
-end if defined?(ActiveSupport::Concern)
+end if defined?(ActiveRecord)